Test BSD, illumos, and arm64 in CI, fix FreeBSD and illumos
Test {Free,Open,Net}BSD, illumos, and {ubuntu,windows}-arm in the CI.

Also update some CI bits and fix FreeBSD; syscall.{Mknod,Mkfifo} have
different types there.

Require go 1.21 because errors.ErrUnsupported was added in 1.21.
