+++ .github/workflows/test.yml
@@ -1,23 +1,87 @@
 name: 'test'
-on:   ['push', 'pull_request']
+on:
+  pull_request:
+    paths: ['**.go', 'go.mod', '.github/workflows/*']
+  push:
+    branches: ['main']
 
 jobs:
-  test:
-    strategy:
-      matrix:
-        go: ['1.20.x', '1.24.x']
-        os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
-    runs-on: ${{ matrix.os }}
-    steps:
-      - uses: 'actions/checkout@v4'
-      - uses: 'actions/setup-go@v5'
-        with:
-          go-version: ${{ matrix.go }}
-      - run:  'go test -race ./...'
-
   staticcheck:
     runs-on: 'ubuntu-latest'
     steps:
-      - uses: 'actions/checkout@v4'
+      - uses: 'actions/checkout@v6'
       - uses: 'dominikh/staticcheck-action@v1.3.1'
         with: {version: '2025.1'}
+
+  test:
+    strategy:
+      matrix:
+        go: ['1.21.x', '1.25.x']
+        os: ['ubuntu-latest', 'ubuntu-24.04-arm', 'macos-latest', 'windows-latest', 'windows-11-arm']
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: 'actions/checkout@v6'
+      - uses: 'actions/setup-go@v6'
+        with:
+          go-version: ${{ matrix.go }}
+      - shell: 'bash'
+        run: |
+          # "-race is not supported on windows/arm64"
+          race='-race'
+          [ $RUNNER_OS = 'Windows' ] && [ $RUNNER_ARCH = 'ARM64' ] && race=''
+
+          go test $race ./...
+
+  freebsd:
+    runs-on: 'ubuntu-latest'
+    name:    'test (freebsd)'
+    timeout-minutes: 10
+    steps:
+      - uses: 'actions/checkout@v6'
+      - name: 'test (freebsd)'
+        id:   'freebsd'
+        uses: 'vmactions/freebsd-vm@v1'
+        with:
+          prepare: 'pkg update -f && pkg install -y go'
+          run:     'go test -race ./...'
+
+  openbsd:
+    runs-on: 'ubuntu-latest'
+    name:    'test (openbsd)'
+    timeout-minutes: 10
+    steps:
+      - uses: 'actions/checkout@v6'
+      - name: 'test (openbsd)'
+        id:   'openbsd'
+        uses: 'vmactions/openbsd-vm@v1'
+        with:
+          prepare: 'pkg_add go'
+          # "-race is not supported on openbsd/amd64"
+          run:     'go test ./...'
+
+  netbsd:
+    runs-on: 'ubuntu-latest'
+    name:    'test (netbsd)'
+    timeout-minutes: 10
+    steps:
+      - uses: 'actions/checkout@v6'
+      - name: 'test (netbsd)'
+        id:   'netbsd'
+        uses: 'vmactions/netbsd-vm@v1'
+        with:
+          prepare: 'pkg_add go'
+          run:     '/usr/pkg/bin/go??? test -race ./...'
+
+  illumos:
+    runs-on: 'ubuntu-latest'
+    name:    'test (illumos)'
+    timeout-minutes: 10
+    steps:
+    - uses: 'actions/checkout@v6'
+    - name: 'test (illumos)'
+      id:   'illumos'
+      uses: 'vmactions/omnios-vm@v1'
+      with:
+        prepare: 'pkg install go-123'
+        # "-race is not supported on illumos/amd64"
+        run:     'go test ./...'
