+++ go.mod
@@ -1,3 +1,3 @@
 module zgo.at/zstd
 
-go 1.20
+go 1.21
