+++ ztest/shell.go
@@ -193,3 +193,27 @@ func Truncate(t *testing.T, sz int64, path ...string) {
 		t.Fatalf("ztest.Truncate(%q): %s", join(path...), err)
 	}
 }
+
+// mkfifo
+func Mkfifo(t *testing.T, path ...string) {
+	t.Helper()
+	if len(path) < 1 {
+		t.Fatalf("ztest.Mkfifo: path must have at least one element: %s", path)
+	}
+	err := mkfifo(join(path...), 0o644)
+	if err != nil {
+		t.Fatalf("ztest.Mkfifo(%q): %s", join(path...), err)
+	}
+}
+
+// mknod
+func Mknod(t *testing.T, dev int, path ...string) {
+	t.Helper()
+	if len(path) < 1 {
+		t.Fatalf("ztest.Mknod: path must have at least one element: %s", path)
+	}
+	err := mknod(join(path...), 0o644, dev)
+	if err != nil {
+		t.Fatalf("ztest.Mknod(%d, %q): %s", dev, join(path...), err)
+	}
+}
