+++ ztime/time_test.go
@@ -8,7 +8,8 @@ import (
 
 func TestTakes(t *testing.T) {
 	tt := Takes(func() { time.Sleep(50 * time.Millisecond) })
-	if tt < 50*time.Millisecond || tt > 60*time.Millisecond {
+	// Large margin because CI can be slow.
+	if tt < 50*time.Millisecond || tt > 70*time.Millisecond {
 		t.Error(tt)
 	}
 }
